Add files via upload
Develop a  program that uses while statement to determine the gross pay for each of several employees.  The company pays “straight time” for the first 40 hours 	worked by each employee and pays “time-and-a-half” for all hours worked in excess of 40 hours.  Your program should input this information for each employee and should determine and display the employee’s gross pay.

EXAMPLE RUN:
ENTER HOURS WORKED(-1 TO END): 39
ENTER HOURLY RATE OF THE WORKERS($00.00): 10.00
THE SALARY IS: $390.00

ENTER HOURS WORKED(-1 TO END): 40
ENTER HOURLY RATE OF THE WORKERS($00.00): 10.00
THE SALARY IS: $400.00

ENTER HOURS WORKED(-1 TO END): 41
ENTER HOURLY RATE OF THE WORKERS($00.00): 10.00
THE SALARY IS: $415.00

ENTER HOURS WORKED(-1 TO END): -1

THE TOTAL SALARY OF ALL EMPLOYEES: $1205.00
GOOD DAY! THANK YOU FOR USING THE SYSTEM!